finished user controllers

+++ controllers/userController.js
@@ -1,4 +1,4 @@
-const {User} = require ('../models')
+const {User, Thought} = require ('../models')
 
 module.exports = {
     getUsers(req, res) {
@@ -35,8 +35,39 @@ module.exports = {
         .catch((err) => res.status(500).json(err))
     },
     deleteUser(req, res) {
-        User.findOneAndDelete(
-            {_id: req.params.userId}
+        User.findOneAndDelete({_id: req.params.userId})
+            .then((userData) => 
+                !userData
+                    ? res.status(404).json({message: 'No user with that ID'})
+                    : Thought.deleteMany({_id: {$in: userData.thoughts}})
+            )
+            .then(() => res.json({message: 'User and associated thoughts deleted!'}))
+            .catch((err) => res.status(500).json(err))
+    },
+    addFriend(req, res) {
+        User.findOneAndUpdate(
+            {_id: req.params.userId},
+            {$addToSet: {friends: req.params.friendId}},
+            {runValidators: true, new: true}
         )
+        .then((userData) =>
+            !userData
+                ? res.status(404).json({message: 'No user with that ID'})
+                : res.json(userData)
+        )
+        .catch((err) => res.status(500).json(err))
+    },
+    deleteFriend(req, res) {
+        User.findOneAndUpdate(
+            {_id: req.params.userId},
+            {$pull : {friends: req.params.friendId}},
+            {runValidators: true, new: true}
+        )
+        .then((userData) =>
+            !userData
+                ? res.status(404).json({message: 'No user with that ID'})
+                : res.json(userData)
+        )
+        .catch((err) => res.status(500).json(err))
     }
 }
